backend/requirements.txt
fastapi==0.104.1
bcrypt==4.0.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
pydantic[email]==2.5.2
python-dotenv==1.0.0
pillow==10.1.0
